experiment with runtime.dll for relocations

+++ src/programs/windows/win_dynamic/win_dynamic_lib.h
@@ -2,9 +2,9 @@
 
 #include <stdint.h>
 
-extern uint64_t lib_var_bss;
-extern uint64_t lib_var_data;
+extern uint32_t lib_var_bss;
+extern uint32_t lib_var_data;
 
-uint64_t *get_lib_var_bss();
+uint32_t *get_lib_var_bss();
 
-uint64_t *get_lib_var_data();
+uint32_t *get_lib_var_data();
